Fix bug in MasterLockKeeper
The thread that keeps the master lock died because there was no catch
all, and the exception was allowed to crash the thread. Added catch all
to ensure the thread does not die.

Also fixed the check for if the master lock is present, by ensuring that
we have updated the lock within the last 5 minutes.